+++ .env
@@ -1,0 +1,2 @@
+# REACT_APP_API_KEY = AIzaSyBBSsplX6QcyTaV5f4iIiLYMwcAY0EQkPc
+REACT_APP_API_KEY = AIzaSyB0igz-R1JVgCFDWkkWjtVYD2-bJzyKxJc
