+++ frontend/mapcomponent.js
@@ -1,0 +1,16 @@
+// import MapComponent from './MapComponent';
+import React, { Component } from 'react';
+
+export class MapComponent extends Component {
+
+  render() {
+    return (
+      <div>
+        <h1>Hi! Here is a map component hello hello: </h1> <br />
+        {/* <MapComponent /> */}
+      </div>
+    );
+  }
+}
+
+export default MapComponent;
